-Airbnb style guide configuration using eslint init -Move index.html file to public folder -Remove dist folder in client
