+++ README.md
@@ -1,2 +1,41 @@
+# IKEA PRODUCT CAROUSEL
 # djason-ikea-similar-products
-Reproduce the ikea similar product section https://www.ikea.com/us/en/p/friheten-sleeper-sectional-3-seat-w-storage-skiftebo-dark-gray-s69216757/
+
+>  Reproduce the ikea similar product section https://www.ikea.com/us/en/p/friheten-sleeper-sectional-3-seat-w-storage-skiftebo-dark-gray-s69216757/
+
+
+## Related Projects
+
+  - https://github.com/FEC-HRR41-PICOLLO/djason-ikea-similar-products
+  - https://github.com/FEC-HRR41-PICOLLO/darjama-ikea-reviews/
+  - https://github.com/FEC-HRR41-PICOLLO/ryuuohlee-ikea-item-description
+  - https://github.com/FEC-HRR41-PICOLLO/DavidFuent1-Ikea-PhotoCarousel/
+
+## Table of Contents
+
+1. [Usage](#Usage)
+1. [Requirements](#requirements)
+1. [Development](#development)
+
+## Usage
+
+> Some usage instructions
+
+## Requirements
+
+An `nvmrc` file is included if using [nvm](https://github.com/creationix/nvm).
+
+- Node 6.13.0
+- etc
+
+## Development
+
+### Installing Dependencies
+
+From within the root directory:
+
+```sh
+npm install -g webpack
+npm install
+```
+
