--- client/src/index.jsx
@@ -1,128 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-// import $ from "jquery";
-var $ = require("jquery");
-import ItemDetails from './components/ItemDetails.jsx';
-import 'bootstrap'
-class App extends React.Component{
-  constructor(props){
-    super(props);
-    this.state = {
-      data: [],
-      currentPos: 0,
-      dataAlso: [],
-      currentPosAlso: 0
-    }
-    //console.log = console.warn = console.error = () => {};
-    this.left= this.left;
-    this.rigth= this.rigth;
-    this.mousehover = this.mousehover;
-    this.mouseout = this.mouseout;
-  }
-  /*
-  GET CURRENT URL->
-  window.location.href, document.baseURI, document.URL
-  */
-
-  componentDidMount(){
-    this.getData('')
-    setTimeout(this.getData('Also'), 500);
-  }
-
-  getData(arg){
-    var url = `${window.location.origin}/products/similar/`
-    if(arg!== ''){
-      url = `${window.location.origin}/products/alsolike/`;
-    }
-    $.ajax({
-      url: url,
-      type: 'GET',
-      success: (data)=>{
-        var arrData = []
-        var groupData = []
-        var cn = 0;
-        if(data.error !== null){
-          return ;
-        }else{
-          data.result.forEach((item, index)=>{
-            if(cn ===3){
-              groupData.push(item)
-              arrData.push(groupData)
-              cn=0;
-              groupData = [];
-            }else{
-              groupData.push(item)
-              cn++
-            }
-          })
-        }
-
-        var obj = {}
-        obj[`data${arg}`] = arrData;
-        obj[`currentPos${arg}`] = 0;
-        this.setState(state=>(obj))
-      }
-    })
-  }
-
-  render(){
-    return (
-      <div className="container">
-        <hr style={{width:'90%'}}></hr>
-{/*SIMILAR PRODUCT SECTION START*/}
-        <h3 className='dj-title-section'>Similar products</h3>
-        <div className="row no-gutters carousel slide" id="similar" data-ride="carousel" data-interval="false">
-{/* LEFT ARROW */}
-          <div className="col align-self-center">
-            <a className="carousel-control-prev" href="#similar" data-slide="prev">
-              <span className="carousel-control-prev-icon-arrow sp dj-rounded align-self-center" style={{textAlign: 'center'}}>
-                <i className="fas fa-arrow-left" style={{padding:'0.9rem', fontSize:'1.5rem', width:'50px', height:'50px',color:'#000', backgroundColor: 'transparent', borderRadius:'50px'}}></i>
-              </span>
-            </a>
-          </div>
-{/*CAROUSEL ITEM PRODUCT */}
-          <div className="col-10 align-self-center">
-            <ItemDetails listdata={this.state.data} />
-          </div>
-{/* RIGHT ARROW */}
-          <div className="col align-self-center">
-            <a className="carousel-control-next text-center" href="#similar" data-slide="next">
-              <span className="carousel-control-next-icon-arrow sp dj-rounded text-center">
-                  <i className="fas fa-arrow-right dj-icon" style={{padding:'0.9rem', fontSize:'1.5rem', width:'50px', height:'50px',color:'#000', backgroundColor: 'transparent', borderRadius:'50px'}}></i>
-              </span>
-            </a>
-          </div>
-        </div>
-{/*SIMILAR PRODUCT SECTION START*/}
-        <hr style={{width:'90%'}}></hr>
-{/*ALSO LIKE SECTION START*/}
-        <h3 className='dj-title-section'>You might also like</h3>
-        <div className="row no-gutters carousel slide" id="alsolike" data-ride="carousel" data-interval="false">
-{/* LEFT ARROW */}
-          <div className="col align-self-center">
-            <a className="carousel-control-prev" href="#alsolike" data-slide="prev">
-              <span className="carousel-control-prev-icon-arrow sp dj-rounded align-self-center">
-                <i className="fas fa-arrow-left" style={{padding:'0.9rem', fontSize:'1.5rem', width:'50px', height:'50px',color:'#000', backgroundColor: 'transparent', borderRadius:'50px'}}></i>
-              </span>
-            </a>
-          </div>
-{/*CAROUSEL ITEM ALSO LIKE*/}
-          <div className="col-10 align-self-center">
-            <ItemDetails listdata={this.state.dataAlso} />
-          </div>
-{/* RIGHT ARROW */}
-          <div className="col align-self-center">
-            <a className="carousel-control-next text-center" href="#alsolike" data-slide="next">
-              <span className="carousel-control-next-icon-arrow sp dj-rounded text-center">
-                  <i className="fas fa-arrow-right" style={{padding:'0.9rem', fontSize:'1.5rem', width:'50px', height:'50px',color:'#000', backgroundColor: 'transparent', borderRadius:'50px'}}></i>
-              </span>
-            </a>
-          </div>
-        </div>
-{/*ALSO LIKE SECTION END*/}
-      </div>
-    )
-  }
-}
-
-ReactDOM.render(<App />, document.getElementById('app'));
