+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/FEC-HRR41-PICOLLO/djason-ikea-similar-products#readme",
   "dependencies": {
     "babel": "^6.23.0",
+    "body-parser": "^1.19.0",
     "bootstrap": "^4.3.1",
     "cors": "^2.8.5",
     "express": "^4.17.1",
@@ -37,7 +38,13 @@
     "babel-loader": "^6.3.2",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.23.0",
-    "webpack": "^2.2.1",
-    "jest": "^24.9.0"
+    "eslint": "^6.5.1",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.16.0",
+    "eslint-plugin-react-hooks": "^1.7.0",
+    "jest": "^24.9.0",
+    "webpack": "^2.2.1"
   }
 }
