+++ public/css/style.css
@@ -1,0 +1,103 @@
+/* BADGE C*/
+.dj-new-alert{
+  -webkit-transform: translateY(-100%);
+  -ms-transform: translateY(-100%);
+  transform: translateY(-100%);
+  position: absolute;
+  font-weight: 700;
+  color: #fff;
+  background: #ca5007;
+  font-size: .75rem;
+  line-height: 1rem;
+  display: inline-block;
+  padding: 0 .3rem .05rem;
+  border-radius: .8rem;
+}
+
+.dj-title-prod{
+  text-transform: uppercase;
+  font-size: 1rem;
+  line-height: 1.5;
+  font-weight: 700;
+}
+
+.dj-desc-prod{
+  font-size: .875rem;
+  line-height: 1rem;
+  height: 20px;
+}
+
+.dj-price-prod{
+  display: inline-block;
+  font-weight: 700;
+  font-size: 1.5rem;
+  line-height: 1.25;
+}
+
+.dj-rating-block{
+  display: block;
+}
+
+.dj-star-prod{
+  display: inline-block;
+  height: 1.2rem;
+  width: 100%;
+  background-size: 1.2rem;
+  position: absolute;
+}
+
+.dj-starcount-prod{
+  color: #111;
+}
+
+.dj-reviewcount-prod{
+  display: inline-block;
+  white-space: nowrap;
+}
+
+.dj-diclaimer-prod{
+  margin-top: .4rem;
+  color: #666;
+}
+
+.no-margin{
+  margin-top: -40px;
+}
+
+.dj-title-section{
+  text-align: center;
+  letter-spacing: -.5px;
+  font-size: 1.5rem;
+  line-height: 1.3334;
+  letter-spacing: -.42px;
+  font-weight: 700;
+  margin-bottom: 2.1rem;
+}
+
+.dj-center-vertical{
+  display: flex;
+  align-items: center;
+}
+
+.col-md-3{
+  display: inline-block;
+}
+
+.col-md-3 img{
+  width:100%;
+  height:auto;
+}
+
+/* ARROW NEXT AND PREVIOUS CONFIGURATION */
+.dj-rounded{
+  background-color:#fff;
+  border-radius: 50px;
+  width: 200px;
+  height: 50px;
+  display: inline-block;
+  text-align: left;
+}
+
+.dj-rounded:hover{
+  background-color:#f2f2f2;
+}
